reorganised modelling tests
